services/summarizer/requirements.txt
fastapi
uvicorn
httpx
supabase
python-dotenv
openai-whisper
groq 